% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{corn.yield}
\alias{corn.yield}
\title{Corn Grain Yield - Dataset}
\format{A data frame with 80 rows and 6 columns.}
\usage{
data(corn.yield)
}
\description{
Data from an experiment investigating the effect of cultivar and fertilizer treatments on corn grain yield.
}
\details{
\itemize{
\item year. Growing Season (Y2016) - factor.
\item field. Field in which the experiment was conducted (F1 and F2) - factor.
\item block. Subdivision within fields providing replications for all treatment combinations - factor.
\item cultivar. Corn cultivars (C1 and C2) - factor.
\item fertilizer. Fertilizer treatments (Control - non treated; P1-D1, P1-D2, P2-D1, and P2-D2)
\item yield. Corn grain yield [bu/ac] adjusted to 15.5\% moisture - numeric.
}
}
