% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TampaWorkshop-package.r
\docType{package}
\name{TampaWorkshop}
\alias{TampaWorkshop}
\alias{TampaWorkshop-package}
\title{TampaWorkshop.}
\description{
This package was created as an example to be used during the *R You Ready for New Tools to Manage Research Resources?* professional workshop held at the annual international ASA-CSSA-SSSA conference in Tampa, FL (Oct 22, 2017).
}
