% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{YieldConversion}
\alias{YieldConversion}
\title{Convert Corn Yield Data from US to SI Units}
\usage{
YieldConversion(x)
}
\arguments{
\item{x}{unique numeric value or list of values to be converted into.}
}
\value{
Converted numeric value or list of values.
}
\description{
Converts corn yield data from bushels per acre to tons per hectare.
}
\examples{
YieldConversion(150)
YieldConversion(seq(from = 100, to = 200, by = 10))
YieldConversion(data$yield.in.bu.per.ac)
}
